+++ PokeApp/app/src/main/java/com/tzikin/pokeapp/domain/GetFavoriteUseCase.kt
@@ -7,7 +7,7 @@ class GetFavoriteUseCase @Inject constructor(
     private val favoriteRepository: FavoriteRepository
 ) {
 
-    suspend operator fun invoke(id: Int): Boolean {
+    suspend operator fun invoke(id: Long): Boolean {
        val value = favoriteRepository.isFavorite(id)
         return value?.isFavorite ?: false
     }
